- core: removes UnitOn and UnitFuncOn (considered as duplicate). + core: rewires the current On and FuncOn to the MultiOn base implementation. Refactors all the implementations into a single one. = docs: minor fixes over the documentation.
